LayoutTests/fast/selectors/lang-extended-filtering-expected.txt

Verify selector specifying extended filetring of :lang() pseudo class

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll(":lang(en)").length is 3
PASS document.querySelectorAll(":lang(en-)").length is 2
PASS document.querySelectorAll(":lang(en--)").length is 1
PASS document.querySelectorAll(":lang(en---)").length is 0

PASS document.querySelectorAll(":lang(de-de)").length is 8
PASS document.querySelectorAll(":lang(de-DE)").length is 8
PASS document.querySelectorAll(":lang(DE-de)").length is 8
PASS document.querySelectorAll(":lang(dE-dE)").length is 8
PASS document.querySelectorAll(":lang(De-De)").length is 8
PASS document.querySelectorAll(":lang(DE-DE)").length is 8
PASS document.querySelectorAll(":lang(Latn)").length is 1
PASS document.querySelectorAll(":lang(Latf)").length is 1
PASS document.querySelectorAll(":lang(Latn-de)").length is 1
PASS document.querySelectorAll(":lang(Latf-de)").length is 1
PASS document.querySelectorAll(":lang(Latn, de-Latn)").length is 3
PASS document.querySelectorAll(":lang(Latf, de-Latf)").length is 2
PASS document.querySelectorAll(":lang(Latn, de-Latn)").length is 3
PASS document.querySelectorAll(":lang(Latf, de-Latf)").length is 2
PASS document.querySelectorAll(":lang(de-DE--, de-DE--)").length is 0

PASS getComputedStyle(document.querySelectorAll(":lang(de-DE, DE-de)")[0]).color is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll(":lang(de-DE, DE-de)")[1]).color is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll(":lang(de-DE, DE-de)")[2]).color is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll(":lang(de-DE, DE-de)")[3]).color is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll(":lang(de-DE, DE-de)")[4]).color is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll(":lang(de-DE, DE-de)")[5]).color is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll(":lang(de-DE, DE-de)")[6]).color is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll(":lang(de-DE, DE-de)")[7]).color is "rgb(1, 2, 3)"

PASS getComputedStyle(document.querySelector(":lang(Latn, Latn-de)")).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelector(":lang(Latf, Latf-de)")).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelector(":lang(Latf, Latn)")).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelector(":lang(Latf, Latn)")).color is "rgb(0, 0, 0)"

PASS document.querySelectorAll(":lang(-en)").length is 0
PASS document.querySelectorAll(":lang(-en-)").length is 0
PASS document.querySelectorAll(":lang(-en--)").length is 0
PASS document.querySelectorAll(":lang(-en, -en-)").length is 0
PASS document.querySelectorAll(":lang(-en-, -en--)").length is 0

PASS document.querySelectorAll(":lang(fr-x)").length is 0
PASS document.querySelectorAll(":lang(fr-xenomorph)").length is 1

PASS document.querySelectorAll(":lang(cocoa-1)").length is 0
PASS document.querySelectorAll(":lang(cocoa-a)").length is 0
PASS document.querySelectorAll(":lang(cocoa-bar)").length is 0
PASS document.querySelectorAll(":lang(cocoa-1, cocoa-bar)").length is 0
PASS document.querySelectorAll(":lang(cocoa-a, cocoa-bar)").length is 0
PASS document.querySelectorAll(":lang(cocoa-1, cocoa-a)").length is 0

PASS document.querySelectorAll(":lang(foo)").length is 1
PASS document.querySelectorAll(":lang(foo-bar)").length is 1
PASS document.querySelectorAll(":lang(foo--bar)").length is 1
PASS document.querySelectorAll(":lang(foo-)").length is 1
PASS document.querySelectorAll(":lang(foo--)").length is 0
PASS document.querySelectorAll(":lang(foo, foo-)").length is 1
PASS document.querySelectorAll(":lang(foo-bar, foo--bar)").length is 1
PASS document.querySelectorAll(":lang(foo-, foo--)").length is 1
PASS document.querySelectorAll(":lang(foo--, foo---)").length is 0
PASS document.querySelectorAll(":lang(id)").length is 1
PASS document.querySelectorAll(":lang(id-)").length is 1
PASS document.querySelectorAll(":lang(id--)").length is 1
PASS document.querySelectorAll(":lang(id---Java)").length is 1
PASS document.querySelectorAll(":lang(id--Java)").length is 1
PASS document.querySelectorAll(":lang(id-Java)").length is 1
PASS document.querySelectorAll(":lang(id---)").length is 0
PASS document.querySelectorAll(":lang(id, id-)").length is 1
PASS document.querySelectorAll(":lang(id-, id--)").length is 1
PASS document.querySelectorAll(":lang(id--Java, id---Java)").length is 1
PASS document.querySelectorAll(":lang(id-Java, id--Java)").length is 1
PASS document.querySelectorAll(":lang(id---, id----)").length is 0

PASS document.querySelectorAll(":lang(tic-tac-toe)").length is 1
PASS document.querySelectorAll(":lang(tic-toe-tac)").length is 0

PASS document.querySelectorAll(":lang(--en--)").length threw exception Error: SyntaxError: DOM Exception 12.
PASS document.querySelectorAll(":lang(---en---)").length threw exception Error: SyntaxError: DOM Exception 12.
PASS document.querySelectorAll(":lang(en us- de- fr-).length") threw exception Error: SyntaxError: DOM Exception 12.
PASS successfullyParsed is true

TEST COMPLETE
